main funcionando, erro na b de ordem 1

+++ avl.h
@@ -22,18 +22,18 @@ typedef struct arvoreAVL {
 ArvoreAVL* criarAVL();
 int vaziaAVL(ArvoreAVL *arv);
 
-int alturaAVL(NoAVL *noAVL, int *esforco);
-int fbAVL(NoAVL *noAVL, int *esforco);
-NoAVL* buscaNoAVL(NoAVL *noAVL, int valor, int *esforco);
-NoAVL* noTrocaAVL(NoAVL *raiz, int *esforco);
+int alturaAVL(NoAVL *noAVL, long long *esforco);
+int fbAVL(NoAVL *noAVL, long long *esforco);
+NoAVL* buscaNoAVL(NoAVL *noAVL, int valor, long long *esforco);
+NoAVL* noTrocaAVL(NoAVL *raiz, long long *esforco);
 
-NoAVL* adicionarAVL(ArvoreAVL *arv, int valor, int *esforco);
-void removerAVL(ArvoreAVL *arv, int valor, int *esforco);
+NoAVL* adicionarAVL(ArvoreAVL *arv, int valor, long long *esforco);
+void removerAVL(ArvoreAVL *arv, int valor, long long *esforco);
 
-void balancearAVL(ArvoreAVL *arv, NoAVL *noAVL, int *esforco);
-NoAVL* rseAVL(NoAVL* noAVL, int *esforco);
-NoAVL* rsdAVL(NoAVL* noAVL, int *esforco);
-NoAVL* rdeAVL(NoAVL* noAVL, int *esforco);
-NoAVL* rddAVL(NoAVL* noAVL, int *esforco);
+void balancearAVL(ArvoreAVL *arv, NoAVL *noAVL, long long *esforco);
+NoAVL* rseAVL(NoAVL* noAVL, long long *esforco);
+NoAVL* rsdAVL(NoAVL* noAVL, long long *esforco);
+NoAVL* rdeAVL(NoAVL* noAVL, long long *esforco);
+NoAVL* rddAVL(NoAVL* noAVL, long long *esforco);
 
 #endif
